Zapret/requirements_build.txt
customtkinter>=5.2.0
pyinstaller>=5.13.0
pillow>=10.0.0



